Additional explaination of overriding properties
I feel that the properties section could use expansion. I understand its syntactic sugar to us but to many kotlin users it is likely to be an important feature.

My use-case was in trying to explain to a colleague why creating a kotlin interface with `fun getThingies(): Thingies` was a bad idea, and that he should instead put `val thingies: Thingies` on his interface. I wanted to refer him to the kotlin lang docs, but was a little disappointed with what is here right now, so I figured I would suggest these changes.

please note: several links will probably need updating,
- this breaks all links to `overriding members`. I think this is worth it because trying to differentiate between methods and properties is likely a disservice. If backwards compatibility is hotly desired it would not be hard to create a 3-# section `overriding members` with two 4-# sections `overriding methods` and `overriding properties`
- https://kotlinlang.org/docs/reference/properties.html#overriding-properties should point to this new section.