docs: add two more sentences about the 'where' clause (#1329)
* Added two more sentences about the 'where' clause

When I read the documentation, I wrongly assumed that a `where` clause created a union type.  If I had looked really carefully, I probably could have figured out that it was an intersection type.  Given how brief the documentation was for this feature, I thought a little clarification would be beneficial.

* Update pages/docs/reference/generics.md

Co-Authored-By: GlenKPeterson <glen@organicdesign.org>